Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
requests~=2.32.3

google-generativeai==0.8.4